Crate tardigrade_grpc

gRPC service wrapper for the Tardigrade runtime powered by tonic.

The wrapper can be used as a building block for gRPC server in the cases when the CLI app is not good enough (e.g., if custom authentication / authorization logic is required).

The generated Protobuf types are not public. This is intentional; the API contract for the gRPC services is not stable yet.

Crate features

At least one of server and client features should be enabled, since the crate is useless otherwise.

server

(On by default)

Exposes server functionality for Tardigrade Runtime and Storage abstractions.

client

(Off by default)

Exposes client functionality to connect to the runtime.

Examples

use tokio::task;
use tonic::transport::Server;

use tardigrade_rt::{
    engine::Wasmtime, runtime::Runtime, TokioScheduler,
    storage::{LocalStorage, Streaming},
};
use tardigrade_grpc::*;

// Build a workflow runtime
let storage = Arc::new(LocalStorage::default());
let (storage, routing_task) = Streaming::new(storage);
task::spawn(routing_task);
let runtime = Runtime::builder(Wasmtime::default(), storage)
    .with_clock(TokioScheduler)
    .build();

// Create services based on the runtime
let service = RuntimeWrapper::new(runtime);
let channels_service = ChannelsServiceServer::new(service.storage_wrapper());
let runtime_service = RuntimeServiceServer::new(service);

// Create gRPC server with the services
let addr: SocketAddr = "[::]:9000".parse()?;
Server::builder()
    .add_service(runtime_service)
    .add_service(channels_service)
    // Add other services and/or configure the server...
    .serve(addr)
    .await?;

Structs

ChannelsServiceServerserver

Service allowing to access and manipulate Tardigrade channels.

Clientclient

gRPC client for the Tardigrade runtime service.

RuntimeServiceServerserver

Tardigrade runtime server interface.

RuntimeWrapperserver

gRPC service wrapper for the Tardigrade runtime.

StorageWrapperserver

gRPC service wrapper for the Tardigrade runtime Storage.

TestServiceServerserver

Test extensions for the Tardigrade runtime.

Constants

SERVICE_DESCRIPTORserver

Serialized file descriptor set for messages and services declared in this crate.

Traits

WithClockTypeserver

Scheduler that can be described.